% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{coef.het_ensemble}
\alias{coef.het_ensemble}
\title{Coefficients for het_ensemble}
\usage{
\method{coef}{het_ensemble}(object, ...)
}
\arguments{
\item{object}{A fitted \code{het_ensemble} object.}

\item{...}{Unused.}
}
\description{
Coefficients for het_ensemble
}
