% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{.analytic_se_ate}
\alias{.analytic_se_ate}
\title{Analytic SE for ATE via influence functions}
\usage{
.analytic_se_ate(object, ...)
}
\description{
Analytic SE for ATE via influence functions
}
\keyword{internal}
