% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.apply_standardization}
\alias{.apply_standardization}
\title{Apply standardization parameters to new data}
\usage{
.apply_standardization(X, center, scale)
}
\arguments{
\item{X}{Numeric matrix}

\item{center}{Vector of means}

\item{scale}{Vector of sds}
}
\value{
Standardized matrix
}
\description{
Apply standardization parameters to new data
}
\keyword{internal}
