% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effects.R
\name{.ate_se}
\alias{.ate_se}
\title{Compute SE for ATE via influence function}
\usage{
.ate_se(object, cate_values, weights = NULL)
}
\description{
Compute SE for ATE via influence function
}
\keyword{internal}
