% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_binary_treatment}
\alias{.check_binary_treatment}
\title{Check that treatment is binary}
\usage{
.check_binary_treatment(treatment)
}
\arguments{
\item{treatment}{Treatment vector}
}
\value{
Logical
}
\description{
Check that treatment is binary
}
\keyword{internal}
