% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.clip}
\alias{.clip}
\title{Clip values to a range}
\usage{
.clip(x, lower = -Inf, upper = Inf)
}
\arguments{
\item{x}{Numeric vector}

\item{lower}{Lower bound}

\item{upper}{Upper bound}
}
\value{
Clipped vector
}
\description{
Clip values to a range
}
\keyword{internal}
