% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{.cv_loss}
\alias{.cv_loss}
\title{Calculate CV loss for a prediction matrix}
\usage{
.cv_loss(cv_predictions, y, family = "gaussian", weights = NULL)
}
\arguments{
\item{cv_predictions}{Matrix of CV predictions (n x n_learners)}

\item{y}{True outcomes}

\item{family}{"binomial" or "gaussian"}

\item{weights}{Observation weights}
}
\value{
Vector of CV losses by learner
}
\description{
Calculate CV loss for a prediction matrix
}
\keyword{internal}
