% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.detect_family}
\alias{.detect_family}
\title{Get default family based on outcome}
\usage{
.detect_family(y)
}
\arguments{
\item{y}{Outcome vector}
}
\value{
Character, "binomial" or "gaussian"
}
\description{
Get default family based on outcome
}
\keyword{internal}
