% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_optimization.R
\name{.ensemble_predict}
\alias{.ensemble_predict}
\title{Compute ensemble predictions given weights}
\usage{
.ensemble_predict(predictions, weights)
}
\arguments{
\item{predictions}{Matrix of predictions (n x K) or list of prediction vectors}

\item{weights}{Named vector of weights}
}
\value{
Vector of ensemble predictions
}
\description{
Compute ensemble predictions given weights
}
\keyword{internal}
