% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{.fit_elastic_net}
\alias{.fit_elastic_net}
\title{Fit Elastic Net model}
\usage{
.fit_elastic_net(
  X,
  y,
  treatment,
  family = "gaussian",
  weights = NULL,
  alpha = 0.5,
  ...
)
}
\description{
Fit Elastic Net model
}
\keyword{internal}
