% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{.fit_randomforest}
\alias{.fit_randomforest}
\title{Fit Random Forest model}
\usage{
.fit_randomforest(X, y, treatment, family = "gaussian", weights = NULL, ...)
}
\description{
Fit Random Forest model
}
\keyword{internal}
