% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners_modern.R
\name{.get_causal_forest_variance}
\alias{.get_causal_forest_variance}
\title{Get CATE variance from causal forest}
\usage{
.get_causal_forest_variance(model, newX)
}
\description{
Get CATE variance from causal forest
}
\keyword{internal}
