% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_registry.R
\name{.get_learner}
\alias{.get_learner}
\title{Get learner specification by name}
\usage{
.get_learner(name)
}
\arguments{
\item{name}{Learner name}
}
\value{
Learner specification list
}
\description{
Get learner specification by name
}
\keyword{internal}
