% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_registry.R
\name{.get_learners}
\alias{.get_learners}
\title{Get learner specifications for a vector of names}
\usage{
.get_learners(names, check_installed = TRUE)
}
\arguments{
\item{names}{Character vector of learner names}

\item{check_installed}{Check if packages are installed}
}
\value{
Named list of learner specifications
}
\description{
Get learner specifications for a vector of names
}
\keyword{internal}
