% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het_ensemble.R
\name{.handle_propensity}
\alias{.handle_propensity}
\title{Handle propensity score specification}
\usage{
.handle_propensity(propensity, data, treatment, verbose = TRUE)
}
\description{
Handle propensity score specification
}
\keyword{internal}
