% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_registry.R
\name{.init_learner_registry}
\alias{.init_learner_registry}
\title{Initialize the default learner registry}
\usage{
.init_learner_registry()
}
\description{
Initialize the default learner registry
}
\keyword{internal}
