% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_installed}
\alias{.is_installed}
\title{Check if a package is installed}
\usage{
.is_installed(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\value{
Logical
}
\description{
Check if a package is installed
}
\keyword{internal}
