% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_optimization.R
\name{.learner_performance}
\alias{.learner_performance}
\title{Get CV-based performance metrics for learners}
\usage{
.learner_performance(cv_predictions, y, family = "gaussian", weights = NULL)
}
\arguments{
\item{cv_predictions}{Matrix of CV predictions}

\item{y}{True outcomes}

\item{family}{"binomial" or "gaussian"}

\item{weights}{Observation weights}
}
\value{
Data frame with learner performance
}
\description{
Get CV-based performance metrics for learners
}
\keyword{internal}
