% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.logistic}
\alias{.logistic}
\title{Logistic (sigmoid) function}
\usage{
.logistic(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector of probabilities
}
\description{
Logistic (sigmoid) function
}
\keyword{internal}
