% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.logit}
\alias{.logit}
\title{Inverse logistic (logit) function}
\usage{
.logit(p)
}
\arguments{
\item{p}{Numeric vector of probabilities}
}
\value{
Numeric vector
}
\description{
Inverse logistic (logit) function
}
\keyword{internal}
