% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_interaction_matrix}
\alias{.make_interaction_matrix}
\title{Create interaction matrix between covariates and treatment}
\usage{
.make_interaction_matrix(X, treatment, include_main = TRUE)
}
\arguments{
\item{X}{Covariate matrix}

\item{treatment}{Treatment vector}

\item{include_main}{Logical, include main effects}
}
\value{
Interaction matrix
}
\description{
Create interaction matrix between covariates and treatment
}
\keyword{internal}
