% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.message}
\alias{.message}
\title{Create a progress message}
\usage{
.message(msg, verbose = TRUE)
}
\arguments{
\item{msg}{Message}

\item{verbose}{Logical}
}
\description{
Create a progress message
}
\keyword{internal}
