% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_interface.R
\name{.parse_simple_formula}
\alias{.parse_simple_formula}
\title{Parse a simple formula (outcome ~ covariates)}
\usage{
.parse_simple_formula(formula, data)
}
\arguments{
\item{formula}{Formula object}

\item{data}{Data frame}
}
\value{
List with outcome and covariate names
}
\description{
Parse a simple formula (outcome ~ covariates)
}
\keyword{internal}
