% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.predict_batch}
\alias{.predict_batch}
\title{Batch prediction for large datasets}
\usage{
.predict_batch(
  object,
  newdata,
  batch_size = 10000,
  type = "ensemble",
  verbose = TRUE
)
}
\arguments{
\item{object}{het_ensemble object}

\item{newdata}{Data frame}

\item{batch_size}{Number of observations per batch}

\item{type}{Prediction type}

\item{verbose}{Show progress}
}
\value{
Predictions
}
\description{
Predicts in batches to manage memory for large datasets.
}
\keyword{internal}
