% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{.predict_bayesglm}
\alias{.predict_bayesglm}
\title{Predict from Bayesian GLM}
\usage{
.predict_bayesglm(model, newX, treatment = NULL, family = "gaussian")
}
\description{
Predict from Bayesian GLM
}
\keyword{internal}
