% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.predict_cate}
\alias{.predict_cate}
\title{CATE prediction (treatment effect for each observation)}
\usage{
.predict_cate(object, X, X_full = NULL)
}
\description{
CATE prediction (treatment effect for each observation)
}
\keyword{internal}
