% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.predict_counterfactual}
\alias{.predict_counterfactual}
\title{Counterfactual predictions}
\usage{
.predict_counterfactual(object, X, X_full = NULL, treatment = NULL)
}
\description{
Counterfactual predictions
}
\keyword{internal}
