% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{.predict_elastic_net}
\alias{.predict_elastic_net}
\title{Predict from Elastic Net model}
\usage{
.predict_elastic_net(model, newX, treatment = NULL, family = "gaussian")
}
\description{
Predict from Elastic Net model
}
\keyword{internal}
