% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{.predict_findit}
\alias{.predict_findit}
\title{Predict from FindIt model}
\usage{
.predict_findit(model, newX, treatment = NULL, family = "binomial")
}
\description{
Predict from FindIt model
}
\keyword{internal}
