% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{.predict_glmboost}
\alias{.predict_glmboost}
\title{Predict from GLMBoost model}
\usage{
.predict_glmboost(model, newX, treatment = NULL, family = "gaussian")
}
\description{
Predict from GLMBoost model
}
\keyword{internal}
