% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.predict_individual}
\alias{.predict_individual}
\title{Individual learner predictions}
\usage{
.predict_individual(object, X, treatment, X_full = NULL)
}
\description{
Individual learner predictions
}
\keyword{internal}
