% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{.predict_lasso}
\alias{.predict_lasso}
\title{Predict from LASSO model}
\usage{
.predict_lasso(model, newX, treatment = NULL, family = "gaussian")
}
\description{
Predict from LASSO model
}
\keyword{internal}
