% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{.predict_randomforest}
\alias{.predict_randomforest}
\title{Predict from Random Forest model}
\usage{
.predict_randomforest(model, newX, treatment = NULL, family = "gaussian")
}
\description{
Predict from Random Forest model
}
\keyword{internal}
