% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners_modern.R
\name{.predict_xgboost}
\alias{.predict_xgboost}
\title{Predict from XGBoost model}
\usage{
.predict_xgboost(model, newX, treatment = NULL, family = "gaussian")
}
\description{
Predict from XGBoost model
}
\keyword{internal}
