% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.recode_treatment}
\alias{.recode_treatment}
\title{Convert treatment to 0/1 coding}
\usage{
.recode_treatment(treatment, control = NULL, treated = NULL)
}
\arguments{
\item{treatment}{Treatment vector}

\item{control}{Control value}

\item{treated}{Treated value}
}
\value{
Numeric 0/1 vector
}
\description{
Convert treatment to 0/1 coding
}
\keyword{internal}
