% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.require_package}
\alias{.require_package}
\title{Safely load a package}
\usage{
.require_package(pkg, fn = NULL)
}
\arguments{
\item{pkg}{Package name}

\item{fn}{Function name to check}
}
\value{
Logical indicating success
}
\description{
Safely load a package
}
\keyword{internal}
