% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.safe_try}
\alias{.safe_try}
\title{Safe try-catch wrapper}
\usage{
.safe_try(
  expr,
  on_error = c("warn", "stop", "ignore"),
  learner_name = "learner"
)
}
\arguments{
\item{expr}{Expression to evaluate}

\item{on_error}{How to handle errors}

\item{learner_name}{Name of learner (for messages)}
}
\value{
Result or NULL on error
}
\description{
Safe try-catch wrapper
}
\keyword{internal}
