% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.standardize}
\alias{.standardize}
\title{Standardize a numeric vector}
\usage{
.standardize(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Logical, remove NA values}
}
\value{
Standardized vector (mean 0, sd 1)
}
\description{
Standardize a numeric vector
}
\keyword{internal}
