% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.standardize_matrix}
\alias{.standardize_matrix}
\title{Standardize columns of a matrix}
\usage{
.standardize_matrix(X, center = TRUE, scale = TRUE)
}
\arguments{
\item{X}{Numeric matrix}

\item{center}{Logical, center columns}

\item{scale}{Logical, scale columns}
}
\value{
List with standardized matrix and transformation parameters
}
\description{
Standardize columns of a matrix
}
\keyword{internal}
