% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_data}
\alias{.validate_data}
\title{Validate input data}
\usage{
.validate_data(data, outcome, treatment, covariates)
}
\arguments{
\item{data}{Data frame}

\item{outcome}{Outcome column}

\item{treatment}{Treatment column}

\item{covariates}{Covariate columns}
}
\value{
Validated data components
}
\description{
Validate input data
}
\keyword{internal}
