% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.weighted_mean}
\alias{.weighted_mean}
\title{Weighted mean with NA handling}
\usage{
.weighted_mean(x, w = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{w}{Weights}

\item{na.rm}{Remove NAs}
}
\value{
Weighted mean
}
\description{
Weighted mean with NA handling
}
\keyword{internal}
