% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.weighted_var}
\alias{.weighted_var}
\title{Weighted variance}
\usage{
.weighted_var(x, w = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{w}{Weights}

\item{na.rm}{Remove NAs}
}
\value{
Weighted variance
}
\description{
Weighted variance
}
\keyword{internal}
