% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fitted.het_ensemble}
\alias{fitted.het_ensemble}
\title{Get fitted values}
\usage{
\method{fitted}{het_ensemble}(object, type = c("ensemble", "individual"), ...)
}
\arguments{
\item{object}{het_ensemble object}

\item{type}{Type of fitted values}
}
\value{
Vector of fitted values
}
\description{
Get fitted values
}
\keyword{internal}
