% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het_ensemble.R
\name{is.het_ensemble}
\alias{is.het_ensemble}
\title{Check if object is het_ensemble}
\usage{
is.het_ensemble(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if object is het_ensemble
}
\keyword{internal}
