% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effects.R
\name{print.ate}
\alias{print.ate}
\title{Print method for ate}
\usage{
\method{print}{ate}(x, digits = 4, ...)
}
\arguments{
\item{x}{A fitted \code{ate} object.}

\item{digits}{Number of digits to print.}

\item{...}{Unused.}
}
\description{
Print method for ate
}
