% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{print.bootstrap_effects}
\alias{print.bootstrap_effects}
\title{Print method for bootstrap_effects}
\usage{
\method{print}{bootstrap_effects}(x, digits = 4, ...)
}
\arguments{
\item{x}{A fitted \code{bootstrap_effects} object.}

\item{digits}{Number of digits to print.}

\item{...}{Unused.}
}
\description{
Print method for bootstrap_effects
}
