% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{print.het_ensemble}
\alias{print.het_ensemble}
\title{Print het_ensemble}
\usage{
\method{print}{het_ensemble}(x, digits = 4, ...)
}
\arguments{
\item{x}{A fitted \code{het_ensemble} object.}

\item{digits}{Number of digits to print.}

\item{...}{Unused.}
}
\description{
Print het_ensemble
}
