% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effects.R
\name{print.mcate}
\alias{print.mcate}
\title{Print method for mcate}
\usage{
\method{print}{mcate}(x, digits = 4, ...)
}
\arguments{
\item{x}{A fitted \code{mcate} object.}

\item{digits}{Number of digits to print.}

\item{...}{Unused.}
}
\description{
Print method for mcate
}
