% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{print.summary.het_ensemble}
\alias{print.summary.het_ensemble}
\title{Print summary.het_ensemble}
\usage{
\method{print}{summary.het_ensemble}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{summary.het_ensemble} object.}

\item{digits}{Number of digits to print.}

\item{...}{Unused.}
}
\description{
Print summary.het_ensemble
}
