% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{residuals.het_ensemble}
\alias{residuals.het_ensemble}
\title{Get residuals}
\usage{
\method{residuals}{het_ensemble}(object, type = c("response", "deviance"), ...)
}
\arguments{
\item{object}{het_ensemble object}

\item{type}{Type of residuals}
}
\value{
Vector of residuals
}
\description{
Get residuals
}
\keyword{internal}
