% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{summary.het_ensemble}
\alias{summary.het_ensemble}
\title{Summary het_ensemble}
\usage{
\method{summary}{het_ensemble}(object, digits = 4, ...)
}
\arguments{
\item{object}{A fitted \code{het_ensemble} object.}

\item{digits}{Number of digits to print.}

\item{...}{Unused.}
}
\description{
Summary het_ensemble
}
