% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{view_effects}
\alias{view_effects}
\title{Launch Shiny viewer for effects}
\usage{
view_effects(object, ...)
}
\arguments{
\item{object}{A fitted \code{het_ensemble} object.}

\item{...}{Additional arguments passed to \code{shiny::runApp}.}
}
\description{
Launch Shiny viewer for effects
}
